+++ packages/@expo/cli/src/start/interface/interactiveActions.ts
@@ -1,23 +1,19 @@
 import chalk from 'chalk';
 
-import { BLT, printHelp, printItem, printQRCode, printUsage, StartOptions } from './commandsTable';
+import { BLT, StartOptions, printHelp, printItem, printQRCode, printUsage } from './commandsTable';
+import { MoreToolMenuItem, createDevToolsMenuItems } from './createDevtoolMenuItems';
 import * as Log from '../../log';
 import { env } from '../../utils/env';
 import { learnMore } from '../../utils/link';
-import { openBrowserAsync } from '../../utils/open';
-import { ExpoChoice, selectAsync } from '../../utils/prompts';
+import { selectAsync } from '../../utils/prompts';
 import { DevServerManager } from '../server/DevServerManager';
 import {
   openJsInspector,
-  queryAllInspectorAppsAsync,
   promptInspectorAppAsync,
+  queryAllInspectorAppsAsync,
 } from '../server/middleware/inspector/JsInspector';
 
 const debug = require('debug')('expo:start:interface:interactiveActions') as typeof console.log;
-
-interface MoreToolMenuItem extends ExpoChoice<string> {
-  action?: () => unknown;
-}
 
 /** Wraps the DevServerManager and adds an interface for user actions. */
 export class DevServerManagerActions {
@@ -139,25 +135,26 @@ export class DevServerManagerActions {
         { title: 'Toggle performance monitor', value: 'togglePerformanceMonitor' },
         { title: 'Toggle developer menu', value: 'toggleDevMenu' },
         { title: 'Reload app', value: 'reload' },
-        // TODO: Maybe a "View Source" option to open code.
       ];
-      const pluginMenuItems = (
-        await this.devServerManager.devtoolsPluginManager.queryPluginsAsync()
-      ).map((plugin) => ({
-        title: chalk`Open {bold ${plugin.packageName}}`,
-        value: `devtoolsPlugin:${plugin.packageName}`,
-        action: async () => {
-          const url = new URL(
-            plugin.webpageEndpoint,
-            this.devServerManager
-              .getDefaultDevServer()
-              .getUrlCreator()
-              .constructUrl({ scheme: 'http' })
-          );
-          await openBrowserAsync(url.toString());
-        },
-      }));
-      const menuItems = [...defaultMenuItems, ...pluginMenuItems];
+
+      const defaultServerUrl = this.devServerManager
+        .getDefaultDevServer()
+        .getUrlCreator()
+        .constructUrl({ scheme: 'http' });
+
+      const getInspectorApps = async () => {
+        const metroServerOrigin = this.devServerManager
+          .getDefaultDevServer()
+          .getJsInspectorBaseUrl();
+        return await queryAllInspectorAppsAsync(metroServerOrigin);
+      };
+
+      const plugins = await this.devServerManager.devtoolsPluginManager.queryPluginsAsync();
+      const menuItems = [
+        ...defaultMenuItems,
+        ...createDevToolsMenuItems(plugins, defaultServerUrl, getInspectorApps),
+      ];
+
       const value = await selectAsync(chalk`Dev tools {dim (native only)}`, menuItems);
       const menuItem = menuItems.find((item) => item.value === value);
       if (menuItem?.action) {